+++ src/nbt_tag/mod.rs
@@ -8,8 +8,6 @@ use std::io::{self, BufWriter, BufReader};
 use derive_new::new;
 
 use pyo3::prelude::*;
-use pyo3::types::{PyDict, PyList};
-use pyo3::exceptions::{PyTypeError, PyKeyError};
 
 #[cfg(test)]
 mod tests;
@@ -90,7 +88,7 @@ impl NbtTagCompound {
 /// NBT is a tag-based binary format used to store structured data.
 /// Each `NbtTagType` variant corresponds to a different data type
 /// in the NBT specification.
-#[pyclass]
+
 #[derive(Clone, Copy, new,  Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
 pub enum NbtTagType {
     End,
@@ -181,27 +179,6 @@ impl Default for NbtTag {
 }
 
 impl NbtTag {
-
-    //https://github.com/PyO3/pyo3/pull/3582 
-    //ENUM Variants are not supported in PyO3. NbtTag like originally implemented cannot be seen in python
-/*     pub fn new(nbt_type: NbtTagType)-> Self {
-        
-        match nbt_type {
-            NbtTagType::End => NbtTag::End,
-            NbtTagType::Byte => NbtTagByte::new(),
-            NbtTagType::Short => NbtTagShort::new(),
-            NbtTagType::Int => NbtTagInt::new(),
-            NbtTagType::Long => NbtTagLong::new(),
-            NbtTagType::Float => NbtTagFloat::new(),
-            NbtTagType::Double => NbtTagDouble::new(),
-            NbtTagType::ByteArray => NbtTagByteArray::new(),
-            NbtTagType::String => NbtTagString::new(),
-            NbtTagType::List => NbtTagList::new(),
-            NbtTagType::Compound => NbtTagCompound::new(),
-            NbtTagType::IntArray => NbtTagIntArray::new(),
-            NbtTagType::LongArray => NbtTagLongArray::new()
-        }
-    } */
 
     pub fn ty(&self) -> NbtTagType {
         match &self {
@@ -319,245 +296,6 @@ impl NbtTag {
 
 }
 
-#[derive(Clone, Debug)]
-pub struct SerializablePyDict(Py<PyDict>);
-
-impl SerializablePyDict {
-    pub fn get_py_dict(&self) -> &Py<PyDict> {
-        &self.0
-    }
-}
-
-impl IntoPy<Py<PyAny>> for SerializablePyDict {
-    fn into_py(self, py: Python) -> Py<PyAny> {
-        self.0.into_py(py)
-    }
-}
-
-impl ToPyObject for SerializablePyDict {
-    fn to_object(&self, py: Python) -> PyObject {
-        self.0.to_object(py) // Delegate to Py<PyDict>'s implementation
-    }
-}
-
-impl FromPyObject<'_> for SerializablePyDict {
-    fn extract(ob: &'_ PyAny) -> PyResult<Self> {
-        let py_dict: Py<PyDict> = ob.extract()?; // Extract as Py<PyDict> using PyDict's FromPyObject
-        Ok(SerializablePyDict(py_dict)) // Wrap in SerializablePyDict
-    }
-}
-
-impl Serialize for SerializablePyDict {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        Python::with_gil(|py| {
-            let dict: &PyDict = self.0.as_ref(py);
-            let mut rust_dict = HashMap::new();
-
-            for (key, value) in dict.into_iter() {
-                let key_str = key.extract::<String>().map_err(serde::ser::Error::custom)?;
-                let value_str = value.extract::<PyNbtTag>().map_err(serde::ser::Error::custom)?;
-                rust_dict.insert(key_str, value_str);
-            }
-
-            let mut map = serializer.serialize_map(Some(rust_dict.len()))?;
-            for (k, v) in rust_dict {
-                map.serialize_entry(&k, &v.ser_python_dict)?;
-            }
-            map.end()
-        })
-    }
-}
-
-
-#[pyclass(get_all)]
-#[derive(Clone, Debug)]
-pub struct PyNbtTag {
-    //pub nbt_tag: &'a NbtTag,
-    //pub python_dict: Py<PyDict>,
-    pub ser_python_dict: SerializablePyDict
-}
-
-impl PyNbtTag {
-
-    pub fn new(nbt_tag: &NbtTag) -> Self {
-        let python_dict = Self::to_python_dictionary(&nbt_tag);
-        let ser_py_dict = Self::to_ser_python_dictionary(python_dict);
-        Self {
-            //python_dict,
-            ser_python_dict: ser_py_dict
-        }
-    }
-
-    fn to_ser_python_dictionary(py_dict: Py<PyDict>) -> SerializablePyDict {
-        SerializablePyDict(py_dict)
-    }
-
-    fn to_python_dictionary(nbt_tag: & NbtTag) -> Py<PyDict> {
-        
-        Python::with_gil(|py| {
-            let dict: Py<PyDict> = PyDict::new(py).into();
-            // TODO: Get rid of all these unwraps
-
-            match nbt_tag.ty() {
-                NbtTagType::End => {
-
-                    let log_msg = format!("tag_end: Name: {}, Value: {}", "[END]", "[END]");
-                    crate::py_log(log_msg);
-
-                    dict.as_ref(py).set_item("END_TAG", 0).unwrap();
-                    dict
-                },
-                NbtTagType::Byte => {
-                    let tag_byte = nbt_tag.byte().unwrap();
-
-                    let log_msg = format!("tag_byte: Name: {}, Value: {}", tag_byte.name, tag_byte.value);
-                    crate::py_log(log_msg);
-
-                    dict.as_ref(py).set_item(tag_byte.name, tag_byte.value).unwrap();
-                    dict
-
-                },
-                NbtTagType::Short => {
-                    let tag_short = nbt_tag.short().unwrap();
-
-                    let log_msg = format!("tag_short: Name: {}, Value: {}", tag_short.name, tag_short.value);
-                    crate::py_log(log_msg);
-
-
-                    dict.as_ref(py).set_item(tag_short.name, tag_short.value).unwrap();
-                    dict
-
-                },
-                NbtTagType::Int => {
-                    let tag_int = nbt_tag.int().unwrap_or_default(); //error without default.
-
-                    let log_msg = format!("tag_int: Name: {}, Value: {}", tag_int.name, tag_int.value);
-                    crate::py_log(log_msg);
-
-                    dict.as_ref(py).set_item(tag_int.name, tag_int.value).unwrap();
-                    dict
-
-                },
-                NbtTagType::Long => {
-                    let tag_long = nbt_tag.long().unwrap();
-
-                    let log_msg = format!("tag_long: Name: {}, Value: {}", tag_long.name, tag_long.value);
-                    crate::py_log(log_msg);
-
-                    dict.as_ref(py).set_item(tag_long.name, tag_long.value).unwrap();
-                    dict
-
-                },
-                NbtTagType::Float => {
-                    let tag_float = nbt_tag.float().unwrap();
-
-                    let log_msg = format!("tag_float: Name: {}, Value: {}", tag_float.name, tag_float.value);
-                    crate::py_log(log_msg);
-
-
-                    dict.as_ref(py).set_item(tag_float.name, tag_float.value).unwrap();
-                    dict
-
-                },
-                NbtTagType::Double => {
-                    let tag_double = nbt_tag.double().unwrap();
-
-                    let log_msg = format!("tag_double: Name: {}, Value: {}", tag_double.name, tag_double.value);
-                    crate::py_log(log_msg);
-
-                    dict.as_ref(py).set_item(tag_double.name, tag_double.value).unwrap();
-                    dict
-
-                },
-                NbtTagType::ByteArray => {
-                    let tag_byte_array = nbt_tag.byte_array().unwrap();
-
-                    let log_msg = format!("tag_byte_array: Name: {}, Value: {}", tag_byte_array.name, "[Values]");
-                    crate::py_log(log_msg);
-
-                    dict.as_ref(py).set_item(tag_byte_array.name, tag_byte_array.values).unwrap();
-                    dict
-
-                },
-                NbtTagType::String => {
-                    let tag_string = nbt_tag.string().unwrap();
-
-                    let log_msg = format!("tag_string: Name: {}, Value: {}", tag_string.name, tag_string.value);
-                    crate::py_log(log_msg);
-
-                    dict.as_ref(py).set_item(tag_string.name, tag_string.value).unwrap();
-                    dict
-
-                },
-                NbtTagType::List => {
-                    let tag_list = nbt_tag.list().unwrap();
-                    let empty_object_array: &[PyObject] = &[];
-                    let py_list: &PyList = PyList::new(py, empty_object_array);
-
-                    let log_msg = format!("tag_list: Name: {}, Value: {}", tag_list.name, "[NbtTagList]");
-                    crate::py_log(log_msg);
-
-                    //not efficient, i am processind the data two times, but for now make it work
-                    for list_element in &tag_list.values {
-                        let py_list_element = PyNbtTag::new(list_element);
-                        let _ = py_list.append(py_list_element.ser_python_dict);
-
-                        let log_msg = format!("tag_list: parsed");
-                        crate::py_log(log_msg);
-                    }
-
-                    dict.as_ref(py).set_item(tag_list.name, py_list).unwrap();
-                    dict
-
-                },
-                NbtTagType::Compound => {
-                    let tag_compound = nbt_tag.compound().unwrap();
-                    //let empty_object_array: &[PyObject] = &[];
-                    let py_dict: &PyDict = PyDict::new(py);
-
-                    let log_msg = format!("tag_compound: Name: {}, Value: {}", tag_compound.name, "[HashMap]");
-                    crate::py_log(log_msg);
-
-                    for (key, value) in tag_compound.values.iter() {
-                        let py_tag = PyNbtTag::new(value);
-                        let _ = py_dict.set_item(key, py_tag.ser_python_dict);
-
-                        let log_msg = format!("tag_compound_hashmap: Name: {}, Value: {}", key, "[NbtTag]");
-                        //let log_msg = format!("tag_compound_hashmap_tag: Name: {}, Value: {}", key, py_tag.python_dict.get_item(key).unwrap());
-                        crate::py_log(log_msg);
-                    }
-
-                    dict.as_ref(py).set_item(tag_compound.name, py_dict).unwrap();
-                    dict
-
-                },
-                NbtTagType::IntArray => {
-                    let tag_int_array = nbt_tag.int_array().unwrap();
-
-                    let log_msg = format!("tag_int_array: Name: {}, Value: {}", tag_int_array.name, "[Values]");
-                    crate::py_log(log_msg);
-
-                    dict.as_ref(py).set_item(tag_int_array.name, tag_int_array.values).unwrap();
-                    dict
-
-                },
-                NbtTagType::LongArray => {
-                    let tag_long_array = nbt_tag.long_array().unwrap();
-
-                    let log_msg = format!("tag_long_array: Name: {}, Value: {}", tag_long_array.name, "[Values]");
-                    crate::py_log(log_msg);
-
-                    dict.as_ref(py).set_item(tag_long_array.name, tag_long_array.values).unwrap();
-                    dict
-
-                }
-            }
-        })
-    }
-}
 
 #[pyclass]
 #[derive(Clone, new, Debug, Default, Serialize, Deserialize)]
