+++ tests/outputs/output_bt.json
@@ -1,6 +1,12 @@
 {
   "name": "Level",
   "values": {
+    "intTest": {
+      "Int": {
+        "name": "intTest",
+        "value": 2147483647
+      }
+    },
     "byteArrayTest (the first 1000 values of (n*n*255+n*7)%100, starting with n=0 (0, 62, 34, 16, 8, ...))": {
       "ByteArray": {
         "name": "byteArrayTest (the first 1000 values of (n*n*255+n*7)%100, starting with n=0 (0, 62, 34, 16, 8, ...))",
@@ -1008,82 +1014,6 @@
         ]
       }
     },
-    "intTest": {
-      "Int": {
-        "name": "intTest",
-        "value": 2147483647
-      }
-    },
-    "floatTest": {
-      "Float": {
-        "name": "floatTest",
-        "value": 0.49823147
-      }
-    },
-    "listTest (compound)": {
-      "List": {
-        "name": "listTest (compound)",
-        "ty": "Compound",
-        "values": [
-          {
-            "Compound": {
-              "name": "",
-              "values": {
-                "name": {
-                  "String": {
-                    "name": "name",
-                    "value": "Compound tag #0"
-                  }
-                },
-                "created-on": {
-                  "Long": {
-                    "name": "created-on",
-                    "value": 1264099775885
-                  }
-                }
-              }
-            }
-          },
-          {
-            "Compound": {
-              "name": "",
-              "values": {
-                "name": {
-                  "String": {
-                    "name": "name",
-                    "value": "Compound tag #1"
-                  }
-                },
-                "created-on": {
-                  "Long": {
-                    "name": "created-on",
-                    "value": 1264099775885
-                  }
-                }
-              }
-            }
-          }
-        ]
-      }
-    },
-    "shortTest": {
-      "Short": {
-        "name": "shortTest",
-        "value": 32767
-      }
-    },
-    "byteTest": {
-      "Byte": {
-        "name": "byteTest",
-        "value": 127
-      }
-    },
-    "stringTest": {
-      "String": {
-        "name": "stringTest",
-        "value": "HELLO WORLD THIS IS A TEST STRING ÃÃÃ!"
-      }
-    },
     "listTest (long)": {
       "List": {
         "name": "listTest (long)",
@@ -1122,10 +1052,10 @@
         ]
       }
     },
-    "longTest": {
-      "Long": {
-        "name": "longTest",
-        "value": 9223372036854775807
+    "shortTest": {
+      "Short": {
+        "name": "shortTest",
+        "value": 32767
       }
     },
     "nested compound test": {
@@ -1136,16 +1066,16 @@
             "Compound": {
               "name": "ham",
               "values": {
-                "name": {
-                  "String": {
-                    "name": "name",
-                    "value": "Hampus"
-                  }
-                },
                 "value": {
                   "Float": {
                     "name": "value",
                     "value": 0.75
+                  }
+                },
+                "name": {
+                  "String": {
+                    "name": "name",
+                    "value": "Hampus"
                   }
                 }
               }
@@ -1177,6 +1107,76 @@
       "Double": {
         "name": "doubleTest",
         "value": 0.4931287132182315
+      }
+    },
+    "stringTest": {
+      "String": {
+        "name": "stringTest",
+        "value": "HELLO WORLD THIS IS A TEST STRING ÃÃÃ!"
+      }
+    },
+    "longTest": {
+      "Long": {
+        "name": "longTest",
+        "value": 9223372036854775807
+      }
+    },
+    "floatTest": {
+      "Float": {
+        "name": "floatTest",
+        "value": 0.49823147
+      }
+    },
+    "listTest (compound)": {
+      "List": {
+        "name": "listTest (compound)",
+        "ty": "Compound",
+        "values": [
+          {
+            "Compound": {
+              "name": "",
+              "values": {
+                "name": {
+                  "String": {
+                    "name": "name",
+                    "value": "Compound tag #0"
+                  }
+                },
+                "created-on": {
+                  "Long": {
+                    "name": "created-on",
+                    "value": 1264099775885
+                  }
+                }
+              }
+            }
+          },
+          {
+            "Compound": {
+              "name": "",
+              "values": {
+                "created-on": {
+                  "Long": {
+                    "name": "created-on",
+                    "value": 1264099775885
+                  }
+                },
+                "name": {
+                  "String": {
+                    "name": "name",
+                    "value": "Compound tag #1"
+                  }
+                }
+              }
+            }
+          }
+        ]
+      }
+    },
+    "byteTest": {
+      "Byte": {
+        "name": "byteTest",
+        "value": 127
       }
     }
   }
